More app refactor
